requirements.txt
pandas
GitPython
tqdm
PyPDF2
numpy
python-dotenv
requests
sentence-transformers